DIS-59 - Receive Shipment Fulfillment with Short Dated Products Information

See merge request arc-one/rsa/biopro-distribution!14

+++ shipping-service/src/main/java/com/arcone/biopro/distribution/shippingservice/infrastructure/listener/dto/OrderFulfilledMessage.java
@@ -13,7 +13,7 @@ public record OrderFulfilledMessage(
     Long shippingCustomerCode,
 
     Long billingCustomerCode,
-    Long locationCode,
+    Integer locationCode,
 
     String deliveryType,
